refactor(rpc): split Program.cs HTTP plumbing + JSON-RPC dispatch into modules
Pure file split, no behaviour change. Differs from PR #428/#429/#431/#432
in mechanism — Program.cs is a top-level program (no partial-class option),
so this is method-extraction with call-site rewrites. Behaviour preservation
proven via build/test counts + source-body equivalence (with qualifier
normalization).

Program.cs: 1353 → 41 LoC (builder, top-level pipeline chain, app.Run).

New layout:
- Dispatch/RpcDispatcher.cs    — DispatchSingleRequest, IsCirclesMethod, IsProxyAllowed
- Dispatch/RpcHandlers.cs      — Handle* + ParseEventParameters + ParseQueryParameters + ReflectionHandler
- Endpoints/HealthCheckEndpoints.cs — /live, /ready, /health
- Endpoints/DocsEndpoints.cs   — /, /docs (HTML index), /openrpc.json, /openrpc
- Endpoints/WebSocketEndpoints.cs   — /ws, /ws/subscribe + HandleUnifiedWebSocket
- Endpoints/JsonRpcEndpoint.cs — POST / single-request route
- Middleware/BatchRpcMiddleware.cs  — POST / batch (JSON array) middleware
- Json/JsonElementExtensions.cs     — global-namespace utility (preserved)

Test coupling bundled:
- OpenRpcGeneratorTests.Generate_AllDispatchedMethods_ArePresentInOpenRpcSpec
  is a source-code drift guard that greps the dispatch switch. The switch
  moved from Program.cs → Dispatch/RpcDispatcher.cs; the test's resolver
  now searches RpcDispatcher.cs first and falls back to Program.cs (so the
  guard still works on pre-split branches).

Verification (clean rebuild against staging baseline):
- Build: 0 errors / 61 warnings (matches baseline)
- Tests: 147 passed / 0 failed / 113 skipped / 260 total (matches baseline)
- Source-level method body equivalence (whitespace-normalized SHA256 with
  RpcHandlers./RpcDispatcher. qualifier-prefix normalization): 31/31 named
  methods byte-equivalent. Two expression-bodied one-liners (IsCirclesMethod,
  IsProxyAllowed) skipped by the brace-matching regex; manually identical
  modulo static→public access modifier.
